fix(so101): keep IK-Abs env cfg importable without Kit
Loading the Isaac-Stack-Cube-SO101-IK-Abs-v0 env cfg eagerly imported
pxr: stack_ik_abs_env_cfg imports SO101PoseIKActionCfg, whose module
imported the SO101PoseIKAction term, which subclasses
DifferentialInverseKinematicsAction and pulls `from pxr import
UsdPhysics` at module load. This broke the forbidden-import contract
that env cfgs must be constructable before SimulationApp launches, so
test_env_cfg_no_forbidden_imports flagged the task.

Split the runtime term into pose_ik_action_term.py and reference it
from the cfg via a lazy "{DIR}.pose_ik_action_term:SO101PoseIKAction"
class_type string, resolved by cfg.validate() after Kit starts. This
mirrors the stock DifferentialInverseKinematicsActionCfg and keeps the
cfg module pxr-free. Update the unit test to resolve the now-string
class_type.
